OasysAnswers wraps up mapping ARN answers across into Oasys
