+++ src/main/kotlin/uk/gov/justice/digital/assessments/api/AssessmentAnswersDto.kt
@@ -3,7 +3,6 @@ package uk.gov.justice.digital.assessments.api
 import java.util.UUID
 
 data class AssessmentAnswersDto(
-
   val assessmentUuid: UUID,
   // Question Code -> List Of Answer Schema Dto
   val answers: Map<String, Collection<AnswerSchemaDto>>
